Move old code into appropriately named folder
By the end of this ticket it should be deleted for good
