+++ packages/backend/src/old/util/StatusError.ts
