+++ packages/backend/src/old/util/auth.ts
